Add tests for the sync download functionality
Create a mock io device that behaves like a DFU device as well as doing
various sanity checks in how it's driven. This is a precursor to start
changing the state machine to improve compatibility.

This only tests for devices with manifestation_tolerant false as the
current code doesn't handle manifestation_tolerant true correctly.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>