Remove special-casing of dfuManifest state
This drops the special-cases of dfuManifest in WaitState, so it really
only waits for a given target state. This is in preperation for handling
manifest tolerant devices correctly (where dfuManifest is not the
final observed state of a download operation).

As a side-effect also drop the UsbReset command and make it a
alternative final Step to do as part of the download loop instead, which
also seems slightly more fitting.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>