Validate intermediate state while waiting
In the cases where the host needs to wait for a device to reach a
certain state (e.g. DfuDnloadIdle) there is an expected
intermediate state for the device to stay in until that point. For
example after a DFU_DNLOAD the device should return DfuDnbusy while
processing and move to DfuDnloadIdle when finished.

Let WaitStatus sanity check that the device is either in the expected intermediate
state or final state. This allows detecting of an out of sync state
machine or the device going into some error condition

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>