Sjoerd/add tests (#18)
* Add function to pull DfuSync apart

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

* Fix State::DfuDnLoadSync typo

switch State::DfuUnloadSync -> State::DfuDnloadSync; Also bump the major
version as State is a public enum.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

* Implement From/Into u8 for State and Status and Eq

Add From implementations in both direction for u8 from/to State and
Status as well as Eq. This makes writing tests later a bit more
convenient.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

* Add tests for the sync download functionality

Create a mock io device that behaves like a DFU device as well as doing
various sanity checks in how it's driven. This is a precursor to start
changing the state machine to improve compatibility.

This only tests for devices with manifestation_tolerant false as the
current code doesn't handle manifestation_tolerant true correctly.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

* Fix clippy unnecessary_lazy_evaluations warning

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>