Support devices that are manifestation tolerant (#19)
* Test if mock device is reset when expected

The only time at the end of a download the mock should have seen a reset
is if  the device won't detach by itself and isn't manifestation
tolerant. Or more specifically when the state machine is
DfuManifestWaitReset and it's the host that's meant to do the resetting

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

* Remove special-casing of dfuManifest state

This drops the special-cases of dfuManifest in WaitState, so it really
only waits for a given target state. This is in preperation for handling
manifest tolerant devices correctly (where dfuManifest is not the
final observed state of a download operation).

As a side-effect also drop the UsbReset command and make it a
alternative final Step to do as part of the download loop instead, which
also seems slightly more fitting.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

* Validate intermediate state while waiting

In the cases where the host needs to wait for a device to reach a
certain state (e.g. DfuDnloadIdle) there is an expected
intermediate state for the device to stay in until that point. For
example after a DFU_DNLOAD the device should return DfuDnbusy while
processing and move to DfuDnloadIdle when finished.

Let WaitStatus sanity check that the device is either in the expected intermediate
state or final state. This allows detecting of an out of sync state
machine or the device going into some error condition

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

* Correctly support manifest tolerant devices

After finishing downloading for manifest tolerant devices, they should
respond with state DfuManifest (manifesting the firmware). Until the
processing is done and they return to DfuIdle at which point further dfu
operations can be done (e.g. an upload).

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

* Add tests for manifestation tolerant devices

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>