Support both dfu 1.1 and dfuse protocols
Split protocol specific data for both dfu and dfuse into their own enum
variants and handle the differences. The IO implementation is now
responsible for a reference to protocol specific data rather then
providing a memory layout (which is only for dfuse).

To make that a bit simpler for IO implementations the code from dfu-libusb
was integrated to do the differentiation between the protocols and the
parsing of dfuse data.

Also add helpers for sending a detach request as well as a usb reset
without having to pull out the IO objects

Fixes #17

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>