Correct page erasure sizes
In commit 4dc4087f76 the page size for erasure was refactored to be
only determined once however; That failed to take into account that the
page sizes can differ in different zones. So refactor once again to be
closer to the original approach of memory layout tracking.

This also updates the test mock to have two zones for dfuse with
different page sizes and ensures that erasures are always properly done
at page boundaries.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>