`std::thread::sleep` should not be used in `async` context (#28)
`std::thread::sleep` is a blocking function, which should not be used in
an `async` context. Instead, it would be better to consider it as I/O
and add it to the `DfuAsyncIo` trait.

This gives the users the freedom to implement it with whichever `async`
runtime they are using, and it makes `dfu-core` compatible with the
`wasm32-unknown-unknown` target which does not have
`std::thread::sleep`.